backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
PyJWT==2.12.1
gunicorn==23.0.0
